pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanonix-backend"
version = "1.0.0"
description = "Multi-tenant backend platform for university kanoons"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "gunicorn==23.0.0",
  "sqlalchemy[asyncio]==2.0.36",
  "asyncpg==0.30.0",
  "alembic==1.14.0",
  "redis==5.2.1",
  "pydantic-settings==2.7.0",
  "email-validator==2.2.0",
  "python-jose[cryptography]==3.3.0",
  "passlib[argon2]==1.7.4",
  "orjson==3.10.12",
  "structlog==24.4.0",
  "python-multipart==0.0.20",
  "prometheus-fastapi-instrumentator==7.0.0",
  "boto3==1.35.91",
  "aiofiles==24.1.0",
  "httpx==0.28.1"
]

[project.optional-dependencies]
test = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.0",
  "asgi-lifespan==2.1.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
